fixed logfire tests
