+++ tests/cohere/conftest.py
@@ -17,46 +17,6 @@ from mirascope.cohere.tools import CohereTool
 from mirascope.cohere.types import CohereEmbeddingParams
 
 
-class BookTool(CohereTool):
-    title: str
-    author: str
-
-
-@pytest.fixture()
-def fixture_book_tool() -> Type[BookTool]:
-    """Returns the `BookTool` type definition."""
-    return BookTool
-
-
-@pytest.fixture()
-def fixture_cohere_response_with_tools() -> NonStreamedChatResponse:
-    """Returns a Cohere chat response with tools in the response"""
-    return NonStreamedChatResponse(
-        text="test",
-        search_queries=None,
-        search_results=None,
-        documents=None,
-        citations=None,
-        tool_calls=[
-            ToolCall(
-                name="BookTool",
-                parameters={
-                    "title": "The Name of the Wind",
-                    "author": "Patrick Rothfuss",
-                },
-            )
-        ],
-        meta=ApiMeta(billed_units=ApiMetaBilledUnits(input_tokens=1, output_tokens=1)),
-    )
-
-
-@pytest.fixture()
-def fixture_cohere_book_tool() -> Type[BookTool]:
-    """Returns the `BookTool` type definition."""
-
-    return BookTool
-
-
 @pytest.fixture()
 def fixture_cohere_embeddings() -> EmbedResponse_EmbeddingsFloats:
     """Returns a Cohere embeddings response with embedding_types None."""
