+++ tests/cohere/test_calls.py
@@ -58,7 +58,7 @@ def test_cohere_call_call(
 @patch("cohere.Client.chat", new_callable=MagicMock)
 def test_cohere_call_call_with_tools(
     mock_chat: MagicMock,
-    fixture_cohere_book_tool: Type[CohereTool],
+    fixture_book_tool: Type[CohereTool],
     fixture_cohere_response_with_tools: NonStreamedChatResponse,
 ) -> None:
     """Tests that `CohereCall.call` works with tools."""
@@ -68,7 +68,7 @@ def test_cohere_call_call_with_tools(
         prompt_template = ""
         api_key = "test"
 
-        call_params = CohereCallParams(tools=[fixture_cohere_book_tool], model="test")
+        call_params = CohereCallParams(tools=[fixture_book_tool], model="test")
 
     response = TempCall().call()
     assert response.tool is not None
